requirements.txt
Flask==1.1.4
Werkzeug==0.16.0
markupsafe==2.0.1
pynamodb==5.3.2
python-dotenv==0.21.0